"Update AuthControlador, AuthServicio and UsuarioServicio with new login, logout, password change and token deletion functionality; update .env file and remove password from Git tracking"
This commit updates the AuthControlador, AuthServicio and UsuarioServicio files with new functionality for user login, logout, password change and token deletion. It also updates the .env file and removes the password from Git tracking. The commit includes changes to the login, logout, enviarCorreoCambioPass, cambiarPassword, eliminarTokensExpirados and verificarContrasena functions, as well as the addition of the eliminarPermanentemente function. The commit also includes the addition of the PrismaClient and the removal of the connect function from the database file.